+++ addon/-private/execution_context/acceptance.js
@@ -13,5 +13,14 @@ AcceptanceExecutionContext.prototype = {
   click(selector, container) {
     /* global click */
     click(selector, container);
+  },
+
+  fillIn(selector, container, text) {
+    /* global fillIn */
+    if (container) {
+      fillIn(selector, container, text);
+    } else {
+      fillIn(selector, text);
+    }
   }
 };
